+++ DrivingAssistant/DrivingAssistant.AndroidApp/Fragments/SessionFragment.cs
@@ -2,19 +2,25 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Android.App;
+using Android.Content;
 using Android.Graphics;
 using Android.OS;
 using Android.Views;
 using Android.Widget;
+using DrivingAssistant.AndroidApp.Activities;
 using DrivingAssistant.AndroidApp.Adapters.ViewModelAdapters;
 using DrivingAssistant.AndroidApp.Services;
 using DrivingAssistant.Core.Models;
+using Newtonsoft.Json;
 using Fragment = Android.Support.V4.App.Fragment;
 
 namespace DrivingAssistant.AndroidApp.Fragments
 {
     public class SessionFragment : Fragment
     {
+        private readonly User _user;
+
         private ListView _listView;
         private Button _addButton;
         private Button _modifyButton;
@@ -25,13 +31,21 @@ namespace DrivingAssistant.AndroidApp.Fragments
         private View _selectedView;
 
         private SessionService _sessionService;
+        private MediaService _mediaService;
         private ICollection<Session> _currentSessions;
+
+        //============================================================
+        public SessionFragment(User user)
+        {
+            _user = user;
+        }
 
         //============================================================
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
             var view = inflater.Inflate(Resource.Layout.fragment_sessions, container, false);
             _sessionService = new SessionService("http://192.168.100.234:3287");
+            _mediaService = new MediaService("http://192.168.100.234:3287");
             SetupFragmentFields(view);
             SetupListAdapter();
             return view;
@@ -80,7 +94,9 @@ namespace DrivingAssistant.AndroidApp.Fragments
         //============================================================
         private void OnAddButtonClick(object sender, EventArgs e)
         {
-            //TODO
+            var intent = new Intent(Context, typeof(SessionEditActivity));
+            intent.PutExtra("user", JsonConvert.SerializeObject(_user));
+            StartActivity(intent);
         }
 
         //============================================================
@@ -92,12 +108,6 @@ namespace DrivingAssistant.AndroidApp.Fragments
         //============================================================
         private void OnViewMapButtonClick(object sender, EventArgs e)
         {
-
-        }
-
-        //============================================================
-        private async void OnDeleteButtonClick(object sender, EventArgs e)
-        {
             if (_selectedPosition == -1)
             {
                 Toast.MakeText(Context, "No session selected!", ToastLength.Short).Show();
@@ -105,9 +115,40 @@ namespace DrivingAssistant.AndroidApp.Fragments
             }
 
             var session = _currentSessions.ElementAt(_selectedPosition);
-            await _sessionService.DeleteAsync(session.Id);
-            Toast.MakeText(Context, "Session deleted!", ToastLength.Short).Show();
-            await RefreshDataSource();
+            var intent = new Intent(Context, typeof(MapActivity));
+
+            var str = session.StartCoordinates.X + "," + session.StartCoordinates.Y + " " + session.EndCoordinates.X + "," + session.EndCoordinates.Y;
+            intent.PutExtra("points", str);
+            StartActivity(intent);
+        }
+
+        //============================================================
+        private void OnDeleteButtonClick(object sender, EventArgs e)
+        {
+            if (_selectedPosition == -1)
+            {
+                Toast.MakeText(Context, "No session selected!", ToastLength.Short).Show();
+                return;
+            }
+
+            var alert = new AlertDialog.Builder(Context);
+            alert.SetTitle("Confirm Delete");
+            alert.SetMessage("Action cannot be undone");
+            alert.SetPositiveButton("Delete", async (o, args) =>
+            {
+                var session = _currentSessions.ElementAt(_selectedPosition);
+                await _sessionService.DeleteAsync(session.Id);
+                Toast.MakeText(Context, "Session deleted!", ToastLength.Short).Show();
+                await RefreshDataSource();
+            });
+
+            alert.SetNegativeButton("Cancel", (o, args) =>
+            {
+                //NOTHING
+            });
+
+            var dialog = alert.Create();
+            dialog.Show();
         }
     }
 }
